requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
astunparse==1.6.3
attrs==23.2.0
backtrader==1.9.78.123
beautifulsoup4==4.12.3
bitarray==2.9.2
blinker==1.8.2
cbor2==5.6.4
ccxt==4.3.63
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
crytic-compile==0.3.7
cycler==0.12.1
Cython==3.0.10
cytoolz==0.12.3
dateparser==1.2.0
DateTime==5.5
eth-account==0.11.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.1
gunicorn==23.0.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.11.0
hexbytes==0.3.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.4.1
keras-rl==0.4.2
keras-rl2==1.0.5
kiwisolver==1.4.5
libclang==18.1.1
lru-dict==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
mplfinance==0.12.10b0
multidict==6.0.5
namex==0.0.8
norm==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
parsimonious==0.10.0
patsy==0.5.6
pefile==2024.8.26
pillow==10.4.0
playsound==1.3.0
prettytable==3.11.0
protobuf==4.25.3
pycares==4.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.2
pyperclip==1.9.0
pypiwin32==223
PySocks==1.7.1
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
pywin32==306
pywin32-ctypes==0.2.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rlp==4.0.1
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.24.0
setuptools==74.1.1
six==1.16.0
slither-analyzer==0.10.4
sniffio==1.3.1
solc-select==1.0.4
sortedcontainers==2.4.0
soupsieve==2.6
statsmodels==0.14.2
ta==0.11.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
toolz==0.12.1
trio==0.26.2
trio-websocket==0.11.1
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
wcwidth==0.2.13
web3==6.20.3
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.44.0
win10toast==0.9
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.0
yarl==1.9.4
zope.interface==6.4.post2
Flask
gunicorn